requirements.txt
coloredlogs==15.0.1
grpcio-tools==1.60.1
paho-mqtt==1.6.1
pycryptodome==3.20.0
pynq==3.0.1
scipy==1.12.0
websockets==12.0.0
